Allow theme templates to live in app/views
Currently WordPress does not allow us to hook in and add custom templates, and it currently only searches for custom templates down 1 subdirectory level, but 'app/views' is 2 levels deep. This means any custom templates we defined would not show up when get_theme_templates() was called. (Template drop downs in the Page meta box or in an ACF rule set.) So we currently have to add it manually to the WordPress cache, and that requires some nasty hard coding. Hopefully soon it will search either infinite levels or allow us to add templates via the filter.

https://core.trac.wordpress.org/ticket/13265