Set the WordPress Post (actually a TimberPost) object on a view
Provide a shortcut getId() method to get the post/page id. Also do to mixing in the Understory Core trait, accessing the ID attribute of the view will call the getId() function. This allows for seamless functionality in a Trait when trying to get the post meta data, weather the Trait is used in a Custom Post Type Model or a View.